services/trainer/requirements.txt
# Database
psycopg2-binary==2.9.11
pandas==2.2.3

# ML Models
scikit-learn==1.6.1
xgboost==2.1.4

# Experiment Tracking
mlflow==2.21.3

# Utilities
numpy==2.2.3
joblib==1.4.2
